StandardScaler, Ridge vs Lasso Regression
multiple values must be scaled and then
get fit and transformed.
There are two types of Regression for 3+ attributes array: Ridge and Lasso. Find the best alpha value by drawing a graph of log10([1/10000~10000], score). The alpha value where test and train score are the closest is the best alpha value.