Linear Regression, Ridge, Lasso, StandardScaler
First, we can't get scaled array of one row array. Because the formula of scale is (value - mean)/std, in this case, mean is equal to the mean of the only row.
Second, it is recommended to use Ridge and Lasso library when there are more than 2 attributes because Linear Regression has huge underfitting problem.
Third, the predicted value from Ridge and Lasso can be differed by its alpha. The best alpha can be found by comparing the differences between scores of train and test of each alpha.